+++ pkgs/intl4x/lib/src/hook_helpers/hashes.dart
